--HG--
branch : milkyway

+++ lib/PageCamel/Web/ListAndEdit/Files.pm
@@ -36,6 +36,10 @@ sub new($proto, %config) {
         croak("orderby must be defined in " . $self->{modname});
     }
 
+    if(!defined($self->{sizelimit})) {
+        $self->{sizelimit} = 10*1024*1024; # Default limit 10 MB
+    }
+
     return $self;
 }
 
@@ -44,6 +48,7 @@ sub register($self) {
     $self->register_webpath($self->{download}->{webpath} . '/download', "get_download", 'GET');
 
     $self->register_webpath($self->{manage}->{webpath}, "get_manage", 'GET', 'POST');
+    $self->register_continueheader($self->{manage}->{webpath}, "check_continue");
 
     $self->register_webpath($self->{checkfname}->{webpath}, "get_fname", 'POST');
 
@@ -51,6 +56,7 @@ sub register($self) {
 
     if(defined($self->{upload}->{webpath})) {
         $self->register_webpath($self->{upload}->{webpath}, "get_upload", 'PUT', 'DELETE');
+        $self->register_continueheader($self->{upload}->{webpath}, "check_continue");
     }
 
     return;
@@ -75,6 +81,19 @@ sub clean_fname($self, $filename) {
     $filename =~ s/[^$safe_filename_characters]//g;
     return $filename;
 }
+
+sub check_continue($self, $ua) {
+    if(!defined($ua->{headers}->{'Content-Length'})) {
+        return 0;
+    }
+
+    if($ua->{headers}->{'Content-Length'} > $self->{sizelimit}) {
+        return 0;
+    }
+
+    return 1;
+}
+
 
 sub get_manage($self, $ua) {
 
@@ -109,6 +128,10 @@ sub get_manage($self, $ua) {
         #my $fh = $ua->upload("filename");
 
         if($filename ne '' && $realfname ne '' && defined($ua->{files}->{$realfname}->{data})) {
+            if(length($ua->{files}->{$realfname}->{data}) > $self->{sizelimit}) {
+                # Too big
+                return (status => 413); # Content too large
+            }
             # First delete the existing file (if there is one)
             my $delsth = $dbh->prepare_cached("DELETE FROM " . $self->{tablename} . " WHERE filename = ?")
                     or croak($dbh->errstr);
@@ -178,6 +201,10 @@ sub get_upload($self, $ua) {
     }
 
     my $data = $ua->{postdata};
+    if(length($data) > $self->{sizelimit}) {
+        # Too big
+        return (status => 413); # Content too large
+    }
 
     my $description = 'Uploaded via API';
